Major ntJoin code refactoring
* Separated code into several files for better organization
* Separated code that is specific to scaffolding and synteny modes
* Retained a common class Ntjoin (ntjoin.py) which is a superclass for NtjoinSynteny and NtjoinScaffolder,
which extend the functionality for their specific purposes
* Extracted helper classes to their own dedicated files
* Now, the pipeline is launched by the ntjoin_run.py script
