+++ ntJoin
@@ -230,7 +230,7 @@ ifeq ($(overlap), True)
 endif
 	
 $(target).k$(k).w$(w).n$(n).assigned.scaffolds.fa: $(target).k$(k).w$(w).tsv $(min_ref_targets)
-	$(log_time)  $(assemble_path)/bin/ntjoin_assemble.py scaffold -p $(prefix) -n $(n) -s $< -l $(target_weight) \
+	$(log_time)  $(assemble_path)/bin/ntjoin_run.py scaffold -p $(prefix) -n $(n) -s $< -l $(target_weight) \
 	-r "$(reference_weights_list)" -k $(k) -g $(g) -G $(G) -t $(assemble_t) $(ntjoin_assembly_opts) $(min_ref_targets)
 
 $(target).k$(k).w$(w).n$(n).unassigned.scaffolds.fa: $(target).k$(k).w$(w).n$(n).assigned.scaffolds.fa
@@ -257,4 +257,4 @@ endif
 
 # Synteny
 $(prefix).synteny_blocks.tsv: $(min_ref_targets)
-	$(assemble_path)/bin/ntjoin_assemble.py synteny $^ -k$(k) -w$(w) --w-rounds $(w_sizes)
+	$(assemble_path)/bin/ntjoin_run.py synteny $^ -k$(k) -w$(w) --w-rounds $(w_sizes)
